+++ src/main/java/io/coodoo/workhorse/jobengine/boundary/JobEngineService.java
@@ -5,7 +5,6 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.Objects;
 
-import javax.ejb.EJB;
 import javax.ejb.Stateless;
 import javax.ejb.TransactionAttribute;
 import javax.ejb.TransactionAttributeType;
@@ -51,10 +50,10 @@ public class JobEngineService {
     @Inject
     JobEngine jobEngine;
 
-    @EJB
+    @Inject
     JobQueuePoller jobQueuePoller;
 
-    @EJB
+    @Inject
     JobScheduler jobScheduler;
 
     @Inject